+++ flows/events/base_test.go
@@ -223,6 +223,11 @@ func TestEventMarshaling(t *testing.T) {
 					],
 					"date_format": "YYYY-MM-DD",
 					"default_language": "eng",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "hh:mm",
 					"timezone": "America/Guayaquil"