+++ flows/triggers/base_test.go
@@ -58,7 +58,7 @@ func TestTriggerMarshaling(t *testing.T) {
 	sessionAssets, err := engine.NewSessionAssets(source)
 	require.NoError(t, err)
 
-	env := utils.NewDefaultEnvironment()
+	env := utils.NewEnvironmentBuilder().Build()
 	flow := assets.NewFlowReference(assets.FlowUUID("7c37d7e5-6468-4b31-8109-ced2ef8b5ddc"), "Registration")
 	channel := assets.NewChannelReference("3a05eaf5-cb1b-4246-bef1-f277419c83a7", "Nexmo")
 
@@ -86,6 +86,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"
@@ -123,6 +128,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"
@@ -160,6 +170,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"
@@ -200,6 +215,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"
@@ -236,6 +256,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"
@@ -276,6 +301,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"
@@ -309,6 +339,11 @@ func TestTriggerMarshaling(t *testing.T) {
 				},
 				"environment": {
 					"date_format": "YYYY-MM-DD",
+					"max_value_length": 640,
+					"number_format": {
+						"decimal_symbol": ".",
+						"digit_grouping_symbol": ","
+					},
 					"redaction_policy": "none",
 					"time_format": "tt:mm",
 					"timezone": "UTC"