+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+v0.27.3
+----------
+ * Fix migration of weekday() to add 1
+ * Add resthook_called event
+
+v0.27.2
+----------
+ * simplify reading environments from JSON
+ * rename .Environment() to .Build()
+
+v0.27.1
+----------
+ * Add max_value_length to environment and apply in name and field change modifiers
+
 v0.27.0
 ----------
  * Add type to sessions (the type of the flow it was triggered with)